Removing entrypoint and wait-up scripts
This is changing the image from being something which is run as a long-lived container which can be exec'd into, to one which can be run with the working directory mounted and removed once the current job is finished, as discussed.  By keeping state in a mounted directory on the host, a  lot of the complication in the wrapper scripts isn't needed so can be removed.